chore: move connect_with functions to own files
